backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
